Fixing Operation meta-information classes to reflect the revised behavior of JavaCoreCodeWriter.bipush and sipush. Stubbing in StackTrackingTest (however it still needs to be implemented)

+++ net.dougqh.jak/src/net/dougqh/jak/operations/Bipush.java
@@ -38,7 +38,7 @@ public final class Bipush extends Operation {
 	
 	@Override
 	public final Class< ? >[] getStackResultTypes() {
-		return new Class< ? >[] { byte.class };
+		return new Class< ? >[] { int.class };
 	}
 	
 	@Override
